+++ pybamm/models/submodels/interface/sei/total_sei.py
@@ -1,23 +1,19 @@
 #
-# Class summing up contributions to the main (e.g. intercalation) reaction
+# Class summing up contributions to the SEI reaction
 # for cases with primary, secondary, ... reactions e.g. silicon-graphite
 #
 import pybamm
 
 
-class TotalMainKinetics(pybamm.BaseSubModel):
+class TotalSEI(pybamm.BaseSubModel):
     """
-    Class summing up contributions to the main (e.g. intercalation) reaction
+    Class summing up contributions to the SEI reaction
     for cases with primary, secondary, ... reactions e.g. silicon-graphite
 
     Parameters
     ----------
     param :
         model parameters
-    domain : str
-        The domain to implement the model, either: 'Negative' or 'Positive'.
-    reaction : str
-        The name of the reaction being implemented
     options: dict
         A dictionary of options to be passed to the model.
         See :class:`pybamm.BaseBatteryModel`
@@ -25,21 +21,26 @@ class TotalMainKinetics(pybamm.BaseSubModel):
     **Extends:** :class:`pybamm.interface.BaseInterface`
     """
 
-    def __init__(self, param, domain, reaction, options):
-        super().__init__(param, domain, reaction, options=options)
+    def __init__(self, param, options, cracks=False):
+        if cracks is True:
+            self.reaction = "SEI on cracks"
+        else:
+            self.reaction = "SEI"
+        super().__init__(param, options=options)
 
     def get_coupled_variables(self, variables):
-        Domain = self.domain
-        domain = Domain.lower()
-
-        phases = self.options.phases[domain]
+        phases = self.options.phases["negative"]
         # For each of the variables, the variable name without the phase name
         # is constructed by summing all of the variable names with the phases
         for variable_template in [
-            f"{Domain} electrode {{}}volumetric interfacial current density",
-            f"{Domain} electrode {{}}volumetric interfacial current density [A.m-3]",
-            f"X-averaged {domain} electrode {{}}volumetric "
+            f"Negative electrode {{}}{self.reaction} volumetric "
             "interfacial current density",
+            f"Negative electrode {{}}{self.reaction} volumetric "
+            "interfacial current density [A.m-3]",
+            f"X-averaged negative electrode {{}}{self.reaction} volumetric "
+            "interfacial current density",
+            f"Loss of lithium to {{}}{self.reaction} [mol]",
+            f"Loss of capacity to {{}}{self.reaction} [A.h]",
         ]:
             sumvar = sum(
                 variables[variable_template.format(phase + " ")] for phase in phases